+++ .devcontainer/init-database.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+set -e
+
+psql -v ON_ERROR_STOP=1 --username "$POSTGRES_USER" --dbname "$POSTGRES_DB" <<-EOSQL
+    CREATE DATABASE imagineriosearch;
+    CREATE DATABASE imagineriotest;
+EOSQL
+
+echo DONE
